requirements.txt
torch
torchvision
transformers
tqdm
matplotlib
numpy
pillow
